move surface out of geometry.py
